+++ pyproject.toml
@@ -43,10 +43,12 @@ optional-dependencies = {dev = [
 readme = "README.md"
 requires-python = ">=3.11"
 license.file = "LICENCE.md"
-urls.homepage = "https://github.com/UCL-ARC/qw"
 
 [project.scripts]
 qw = "qw.cli:app"
+
+[project.urls]
+homepage = "https://github.com/UCL-ARC/qw"
 
 [tool.coverage]
 report = {skip_covered = true, sort = "cover"}
@@ -76,6 +78,7 @@ version = "0.1.0"
 github3-py = "^4.0.1"
 gitpython = "^3.1.36"
 keyring = "^24.2.0"
+loguru = "^0.7.2"
 python = "^3.8"
 typer = "^0.9.0"
