+++ tests/test_dummy.py
@@ -1,5 +1,5 @@
 """An example set of tests."""
-import qw.qw
+
 
 def test_stupid_example():
     """Test is merely a placeholder."""
